+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/design-system-tracktor/releases)
 
-## v1.1.2
-- Export theme constants
+## v1.1.3
+- Update `TextField` style
+- Update `Select` style
+- Update `Button` style
+- Update `ToggleButton` style